backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
SQLAlchemy
PyMySQL
python-jose[cryptography]
pydantic
pydantic[email]
